* Merging changes from private svn into trunk (up to r38, I am NOT in a hurry to commit 2.4 support, especially with the way people are flaming me at the moment). Changelog: * implement CMSG_BATTLEGROUND_LIST * several warning corrections * disabling character insert queue as I no longer use it, and it hasn't been updated into the new structure anyway * fixed bug where global cooldown would affect spells that it shouldn't. * fixed issue where moving flag would be set wrongly (fixing spells like aimed shot, also) * more debug message cleanup * some notices -> debug * crap debug message removal * fixed speedhack getting triggered on taxi * fixed all cases of incorrect usage of AddItemToFreeSlot and SafeAddItem where there was a possibility of leaking memory * fixed all cases of incorrect usage of AddItemToFreeSlot and SafeAddItem where there was a possibility of double deleting memory * player creation: items in bag slots will now be added into bag slots instead of backpacks. * fix for aura modifiers again + #define SKIP_ALLOCATOR_SHARING 1 * few packet corrections * fixed weird spell values * re-wrote abunch of the player packet wrapper stuff * fixed gossip pois * smsg_item_push_result -> packets.h + added: packets.h, this is so we don't use dynamic (heap) buffers on small packets. + mage - invisibility (20 second effect:P) * fixed mount speed bug when using a bench * don't we love these? (crash fixes) * fixed bug where casting and removing divine spirit would result in spell damage boost * soulbound items should no longer be tradable no matter what * low-risk crash in MapMgr * some crash fixes + event_GetTimeLeft * rewritten speedhack detection system  - cheatengine users won't be able to get around it anymore, since it uses consistant timing  - three strikes and your out, no rooting inbetween + DelaySpeedHack and removed _heartbeatdisable increments * properly implemented SPELL_AURA_FORCE_REACTION, no more weird effects from ashbringer, etc * just in case, we don't want an endless loop here * it should no longer be possible to shapeshift while on a mount * arenas will no longer recreate parties upon starting rated matches * fixed tranquility bug when entering instances * if this doesn't work, screw it * testing * arenas will no longer start with non-full teams * issue 0000040 - falldamage when changing from flight form * [issue 0000074] Madness of the betrayer proc rate -> 50% * [issue 0000087] - Shadowstep * [issue 0000086] - Shadowstep falling damage * [issue 0000089] Warlock - Hellfire corrected * [issue 0000090] Voidwalker sacrifice will now function as expected * [issue 0000079] Lockpicking skill be be incremented by one of zero for saved characters.
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4292 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/ObjectMgr.cpp
@@ -2131,11 +2131,17 @@ void Charter::RemoveSignature(uint32 PlayerGuid)
 
 void Charter::Destroy()
 {
+	if( Slots == 0 )			// ugly hack because of f*cked memory
+		return;
+
 	//meh remove from objmgr
 	objmgr.RemoveCharter(this);
 	// Kill the players with this (in db/offline)
 	CharacterDatabase.Execute( "DELETE FROM charters WHERE charterId = %u", CharterId );
 	Player * p;
+#ifdef WIN32
+	__try {
+#endif
 	for( uint32 i = 0; i < Slots; ++i )
 	{
 		if(!Signatures[i])
@@ -2144,6 +2150,12 @@ void Charter::Destroy()
 		if( p != NULL)
 			p->m_charters[CharterType] = 0;
 	}
+#ifdef WIN32
+	} __except(EXCEPTION_EXECUTE_HANDLER)
+	{
+		return;
+	}
+#endif
 
 	// click, click, boom!
 	delete this;
